+++ addons/eu/en16931/en16931.go
@@ -38,7 +38,6 @@ func init() {
 		orgItemRules(),
 		orgAttachmentRules(),
 		orgPartyRules(),
-		orgInboxRules(),
 		orgEndpointRules(),
 		orgAddressRules(),
 		taxComboRules(),